proxyserver/requirements.txt
flask~=3.1.0
requests~=2.32.3
python-dotenv~=1.1.0
gunicorn~=23.0.0
redis~=6.2.0
sqlalchemy~=2.0.0
flask-sqlalchemy~=3.1.0